Change argsutils.go to a common cmdutils.go
