+++ artifactory/utils/buildutils.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"encoding/base64"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -258,21 +257,4 @@ func (config *BuildInfoConfiguration) SetServerDetails(art auth.ServiceDetails) 
 
 func (config *BuildInfoConfiguration) IsDryRun() bool {
 	return config.DryRun
-}
-
-type BuildConfiguration struct {
-	BuildName   string
-	BuildNumber string
-	Module      string
-	Project     string
-}
-
-func ValidateBuildAndModuleParams(buildConfig *BuildConfiguration) error {
-	if (buildConfig.BuildName == "" && buildConfig.BuildNumber != "") || (buildConfig.BuildName != "" && buildConfig.BuildNumber == "") {
-		return errors.New("the build-name and build-number options cannot be provided separately")
-	}
-	if buildConfig.Module != "" && buildConfig.BuildName == "" && buildConfig.BuildNumber == "" {
-		return errors.New("the build-name and build-number options are mandatory when the module option is provided")
-	}
-	return nil
 }
