+++ artifactory/utils/utils.go
@@ -182,22 +182,6 @@ func CheckIfRepoExists(repository string, artDetails auth.ServiceDetails) error 
 	return nil
 }
 
-// Get build name and number from env, only if both were not provided
-func GetBuildNameAndNumber(buildName, buildNumber string) (string, string) {
-	if buildName != "" || buildNumber != "" {
-		return buildName, buildNumber
-	}
-	return os.Getenv(coreutils.BuildName), os.Getenv(coreutils.BuildNumber)
-}
-
-// Get build project from env, if not provided
-func GetBuildProject(buildProject string) string {
-	if buildProject != "" {
-		return buildProject
-	}
-	return os.Getenv(coreutils.Project)
-}
-
 // This error indicates that the build was scanned by Xray, but Xray found issues with the build.
 // If Xray failed to scan the build, for example due to a networking issue, a regular error should be returned.
 var buildScanError = errors.New("issues found during xray build scan")
