--- common/utils/argsutils.go
@@ -1,92 +1,0 @@
-package utils
-
-import (
-	"errors"
-	"fmt"
-	"strings"
-
-	"github.com/jfrog/jfrog-client-go/utils/errorutils"
-)
-
-// Removes the provided flag and value from the command arguments
-func RemoveFlagFromCommand(args *[]string, flagIndex, flagValueIndex int) {
-	argsCopy := *args
-	// Remove flag from Command if required.
-	if flagIndex != -1 {
-		argsCopy = append(argsCopy[:flagIndex], argsCopy[flagValueIndex+1:]...)
-	}
-	*args = argsCopy
-}
-
-// Find value of required CLI flag in Command.
-// If flag does not exist, the returned index is -1 and nil is returned as the error.
-// Return values:
-// err - error if flag exists but failed to extract its value.
-// flagIndex - index of flagName in Command.
-// flagValueIndex - index in Command in which the value of the flag exists.
-// flagValue - value of flagName.
-func FindFlag(flagName string, args []string) (flagIndex, flagValueIndex int, flagValue string, err error) {
-	flagIndex = -1
-	flagValueIndex = -1
-	for index, arg := range args {
-		// Check current argument.
-		if !strings.HasPrefix(arg, flagName) {
-			continue
-		}
-
-		// Get flag value.
-		flagValue, flagValueIndex, err = getFlagValueAndValueIndex(flagName, args, index)
-		if err != nil {
-			return
-		}
-
-		// If was not the correct flag, continue looking.
-		if flagValueIndex == -1 {
-			continue
-		}
-
-		// Return value.
-		flagIndex = index
-		return
-	}
-
-	// Flag not found.
-	return
-}
-
-// Get the provided flag's value, and the index of the value.
-// Value-index can either be same as flag's index, or the next one.
-// Return error if flag is found, but couldn't extract value.
-// If the provided index doesn't contain the searched flag, return flagIndex = -1.
-func getFlagValueAndValueIndex(flagName string, args []string, flagIndex int) (flagValue string, flagValueIndex int, err error) {
-	indexValue := args[flagIndex]
-
-	// Check if flag is in form '--key=value'
-	indexValue = strings.TrimPrefix(indexValue, flagName)
-	if strings.HasPrefix(indexValue, "=") {
-		if len(indexValue) > 1 {
-			return indexValue[1:], flagIndex, nil
-		}
-		return "", -1, errorutils.CheckError(errors.New(fmt.Sprintf("Flag %s is provided with empty value.", flagName)))
-	}
-
-	// Check if it is a different flag with same prefix, e.g --server-id-another
-	if len(indexValue) > 0 {
-		return "", -1, nil
-	}
-
-	// If reached here, expect the flag value in next argument.
-	if len(args) < flagIndex+2 {
-		// Flag value does not exist.
-		return "", -1, errorutils.CheckError(errors.New(fmt.Sprintf("Failed extracting value of provided flag: %s.", flagName)))
-	}
-
-	nextIndexValue := args[flagIndex+1]
-	// Don't allow next value to be a flag.
-	if strings.HasPrefix(nextIndexValue, "-") {
-		// Flag value does not exist.
-		return "", -1, errorutils.CheckError(errors.New(fmt.Sprintf("Failed extracting value of provided flag: %s.", flagName)))
-	}
-
-	return nextIndexValue, flagIndex + 1, nil
-}
